fix(payroll): vá lỗ hổng logic tính lương
- Trừ nửa ngày công (0.5) nếu nhân viên quên Check-out thay vì tính đủ 1 ngày.
- Đổi kiểu dữ liệu ActualWorkDays từ int sang double ở Payslip & PayslipDto để hỗ trợ công lẻ.
- Khóa trần tổng ngày công (TotalPaidDays) bằng ngày công chuẩn để chặn kẽ hở tự động nhận lương tăng ca khi đi làm cuối tuần.
- Bổ sung Entity Framework Migration (UpdatePayslipWorkDaysToDouble).

+++ aspnet-core/src/QuanLyNhanSu.Application.Contracts/PayslipDto.cs
@@ -10,7 +10,7 @@ namespace QuanLyNhanSu
         public int Month { get; set; }
         public int Year { get; set; }
         public int StandardWorkDays { get; set; }
-        public int ActualWorkDays { get; set; }
+        public double ActualWorkDays { get; set; }
         public int ApprovedLeaveDays { get; set; }
         public int OvertimeDays { get; set; }
         public decimal OvertimePay { get; set; }
